requirements.txt
APScheduler==3.10.1
argon2-cffi==23.1.0
Flask==2.3.3
Flask-Limiter==3.10.0
Flask-Mail==0.10.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.20
flask_compress==1.17
requests==2.32.3
whitenoise==6.6.0
itsdangerous==2.1.2
PyMySQL==1.1.1